Moved ST Reduction list to own file
