+++ packages/protocol/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./blockmodules/AccountStateModule";
 export * from "./blockmodules/BlockHeightHook";
 export * from "./blockmodules/LastStateRootBlockHook";
-export * from "./utils/ProvableHashList";
 export * from "./model/StateTransition";
 export * from "./model/StateTransitionProvableBatch";
 export * from "./model/Option";
@@ -11,11 +10,12 @@ export * from "./model/transaction/SignedTransaction";
 export * from "./model/transaction/RuntimeTransaction";
 export * from "./model/transaction/ValueOption";
 export * from "./model/MethodPublicOutput";
-export * from "./model/StateTransitionReduction";
 export * from "./model/RuntimeLike";
+export * from "./utils/ProvableHashList";
 export * from "./utils/PrefixedProvableHashList";
 export * from "./utils/MinaPrefixedProvableHashList";
 export * from "./utils/ProvableReductionHashList";
+export * from "./utils/StateTransitionReductionList";
 export * from "./utils/utils";
 export * from "./prover/block/BlockProver";
 export * from "./prover/block/BlockProvable";
